+++ CHANGELOG.md
@@ -6,7 +6,9 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
-- `Pane.togglegroup()` builder
+- `children(nodeList)` builder helper to redirect built children to a specific node list (https://github.com/edvin/tornadofx/issues/95) 
+- `buttonbar` builder (https://github.com/edvin/tornadofx/issues/95)
+- `togglegroup` builder
 
 ## [1.4.3] - 2016-04-23